requirements.txt
Django==4.2.13
djangorestframework==3.15.2
pymongo==4.7.3
dnspython==2.6.1
pandas==2.3.3
numpy==2.3.3
openpyxl==3.1.5
python-dotenv==1.0.1
requests==2.32.3
celery==5.4.0
redis==5.0.7
plotly==5.24.1
Pillow>=11.2.0
django-crispy-forms==2.3
crispy-bootstrap5==2024.10